+++ resources/lang/ca/roles.php
@@ -17,5 +17,8 @@ return [
     'assign' => 'Assigna els rols',
     'assign_notify' => 'Ara tens el rol de ',
     'delete' => 'Esborra l\'assignació',
+    'type_info' => 'Els rols per defecte són roles de classe, si vols que els rols siguen de grup, ho pots canviar ací:',
+    'type_class' => 'Rols de classe',
+    'type_group' => 'Rols de grup',
 
 ];