+++ resources/lang/en/roles.php
@@ -17,5 +17,8 @@ return [
     'assign' => 'Assign roles',
     'assign_notify' => 'Now you are the ',
     'delete' => 'Delete the assignation',
+    'type_info' => 'By default, roles are class roles, if you want to make them group roles, you can configure it here:',
+    'type_class' => 'Classroom roles',
+    'type_group' => 'Group roles',
 
 ];