+++ resources/lang/es/roles.php
@@ -17,5 +17,8 @@ return [
     'assign' => 'Asigna los roles',
     'assign_notify' => 'Ahora tienes el rol de ',
     'delete' => 'Elimina la asignación',
+    'type_info' => 'Los roles por defecto son roles de clase, si quieres hacer que los roles sean de grupo, lo puedes configurar aquí:',
+    'type_class' => 'Roles de clase',
+    'type_group' => 'Roles de grupo',
 
 ];